# More numba optimizations