--- a/nbody/numba/nbody/barnes_hut_array/energy.py
+++ b/nbody/numba/nbody/barnes_hut_array/energy.py
@@ -4,6 +4,15 @@ import numpy as np
 from .quadTree import quadArray
 import time
 
+from . import numba_functions
+import numba
+
+@numba.njit
+def compute_force( nbodies, child, center_of_mass, mass, cell_radius, particles, energy):
+    for i in range(particles.shape[0]):
+        acc = numba_functions.computeForce( nbodies, child, center_of_mass, mass, cell_radius, particles[i] )
+        energy[i, 2] = acc[0]
+        energy[i, 3] = acc[1]
 
 def compute_energy(mass, particles, energy):
     print('compute energy:')
@@ -27,9 +36,7 @@ def compute_energy(mass, particles, energy):
 
     print('\tcompute force: ', end='', flush=True)
     t1 = time.time()    
-    for i in range(particles.shape[0]):
-        acc = root.computeForce(particles[i])
-        energy[i, 2:] = acc
+    compute_force( root.nbodies, root.child, root.center_of_mass, root.mass, root.cell_radius, particles, energy )
     energy[:, :2] = particles[:, 2:]
     t2 = time.time()
     print('{:9.4f}ms'.format(1000*(t2-t1)))