Some more numba optimizations.

As an example, times drop from
```
compute energy:
	build tree:       1.2801ms
	compute mass:    40.8738ms
	compute force:   53.6926ms
	total:           96.1435ms
```
to
```
compute energy:
	build tree:       1.3587ms
	compute mass:     0.3927ms
	compute force:   11.3230ms
	total:           13.2930ms
```
and, using 2 threads for the compute force loop:
```
compute energy:
	build tree:       1.3180ms
	compute mass:     0.3901ms
	compute force:    5.5034ms
	total:            7.4086ms
```