@gouarin 2017-11-14
@rolanddenis how is it possible to have a speedup of 13 with only two threads ??