@rolanddenis 2017-11-14
The description is not clear : it is only at the last step that "compute force" step goes from 11.3ms to 5.5ms using 2 threads. Other performance gains come from unrolling and numba-izing functions (see commit descriptions).